+++ src/Concerto/TestBundle/Resources/public/angularjs/app/js/controllers/test_runner_controller.js
@@ -5,7 +5,7 @@ testRunner.controller('testRunnerController', [
     function ($scope, $compile) {
 
         $scope.concertoOptions = {};
-        $scope.init = function (platform_url, appUrl, testSlug, testName, params, debug, keepAliveInterval, keepAliveTolerance, existingSessionHash) {
+        $scope.init = function (platform_url, appUrl, testSlug, testName, params, debug, admin, keepAliveInterval, keepAliveTolerance, existingSessionHash) {
 
             $scope.concertoOptions = angular.extend($scope.concertoOptions, {
                 platformUrl: platform_url,
@@ -14,6 +14,7 @@ testRunner.controller('testRunnerController', [
                 testName: testName,
                 params: params,
                 debug: debug,
+                admin: admin,
                 keepAliveInterval: keepAliveInterval,
                 keepAliveTolerance: keepAliveTolerance,
                 existingSessionHash: existingSessionHash